server/src:  modularization
