+++ server/src/middlewares/auth0.js
@@ -1,0 +1,1 @@
+//aqui el midleware de auth0
